Combined datascraper V11
The code fetches the list of countries and their individual page URL and then loop over the list to fetch data for each country. The data are then saved as JSON file in a separate folder. 228 countries, but 223 sets of data stored. Country index 218,219,220,224,225 does not have death data plots, hence data not fetched or stored as JSON.

tqdm module used for trackbar during data scraping.